bootstrap a minimal mcp server with integration test (#3)
Add a minimal MCP server (with no resources, tools, etc.) activated with
the new `--run-stdio-server` option. (This option is an internal option
to be used by forthcoming user easy setup logic The intent is for the
plugin to write a single script to be used by the user with LLM
clients.)

For testing, also add an integration test which starts the MCP server
and uses a MCP client to connect to and interact with it. (Although the
test is minimal as well since the server doesn't support anything yet.)
This provides a good foundation with future work.

The integration test framework was copied from the [OpenTelemetry
plugin](https://github.com/shoalsoft/shoalsoft-pants-opentelemetry-plugin),
but then modified to be a little more reusable (but no refactoring at
this time to share with the OpenTelemetry plugin).