ask the Pants PEX for its version in test framework (#15)
Query the Pants PEX for a specific major/minor version for its version
instead of having the test hard-code the Pants versions. This means that
the test will just discover whatever Pants version is in the relevant
resolve automatically and is one less item to update when updating Pants
versions.